% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CanadianFootball_package.r
\docType{package}
\name{CanadianFootball}
\alias{CanadianFootball}
\alias{CanadianFootball-package}
\title{CanadianFootball.}
\description{
CanadianFootball.
}
\keyword{CFL,}
\keyword{CFLStats}
\keyword{Canadian}
\keyword{Football,}

