% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pythagorean_expectation.R
\name{cfl_py_exp}
\alias{cfl_py_exp}
\title{Pythagorean Expectation in the Canadian Football League}
\usage{
cfl_py_exp(PointsFor, PointsAgainst, exponent = 2.37)
}
\arguments{
\item{PointsFor}{Points scored by a given team}

\item{PointsAgainst}{Points scored against a given team}
}
\value{
numeric
}
\description{
Returns the Pythagorean Expectation for a team
}
\keyword{CFL,}
\keyword{canadian}
\keyword{cflstats}
\keyword{football,}

