% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/games.R
\name{get_all_season_games}
\alias{get_all_season_games}
\title{Returns games played by a CFL team in one season}
\usage{
get_all_season_games(season)
}
\arguments{
\item{season}{The season of play}

\item{team}{Abbreviated team name}
}
\value{
data.frame
}
\description{
Returns games played by a CFL team in one season
}
\examples{
\dontrun{
# All games played in 2014
get_all_season_games(2014)
}
}
\keyword{CFL,}
\keyword{canadian}
\keyword{cflstats}
\keyword{football,}

