% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/games.R
\name{get_team_season_games}
\alias{get_team_season_games}
\title{Returns games played by a CFL team in one season}
\usage{
get_team_season_games(team, season)
}
\arguments{
\item{team}{Abbreviated team name}

\item{season}{The season of play}
}
\value{
data.frame
}
\description{
These functions return games played in the CFL at a high level
}
\examples{
\dontrun{
# Games played by Saskatchewan in 2014
get_team_season_games('SSK', 2014)
}
}
\keyword{CFL,}
\keyword{canadian}
\keyword{cflstats}
\keyword{football,}

