% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/pythagorean_expectation.R
\name{pythagorean_exponent}
\alias{pythagorean_exponent}
\title{Pythagorean Expectation exponent}
\usage{
pythagorean_exponent(sd_PtsFor, mean_PtsFor)
}
\arguments{
\item{sd_PtsFor}{Standard Deviation of the points scored by all teams}

\item{mean_PtsFor}{Mean of the points scored by all teams}
}
\value{
numeric
}
\description{
calculates the exponent parameter to be used in Pythagorean Expectation
}
\keyword{CFL,}
\keyword{canadian}
\keyword{cflstats}
\keyword{football,}

