% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/season_py_exp.R
\name{single_team_py_exp}
\alias{single_team_py_exp}
\title{Team Pythagorean Expectation}
\usage{
single_team_py_exp(data, team)
}
\arguments{
\item{data}{A dataset output from the get_all_season_games() function}

\item{team}{A team name abbreviation}
}
\value{
data.frame
}
\description{
Calculates pythagorean expectation for one team over one season
}
\keyword{CFL,}
\keyword{canadian}
\keyword{cflstats}
\keyword{football,}

